events: Refactor and document macro code
This doesn't change the code much, it mostly reorganizes it in new
modules.
