+++ crates/ruma-macros/src/events.rs
@@ -1,8 +1,7 @@
 //! Methods and types for generating events.
 
+pub mod enums;
 pub mod event;
 pub mod event_content;
 pub mod event_enum;
-pub mod event_parse;
-pub mod event_type;
-mod util;
+pub mod event_enum_from_event;
